+++ SRC/main.cc
@@ -16,7 +16,7 @@ int main(int argc, char* argv[])
 
 	else if (argc == 2 || argc == 3)
 	{
-		short int port = atoi(argv[1]);
+		unsigned int port = atoi(argv[1]);
 		if( port < 1024 || port > 65535){
 			std::cerr << "Ça va pas non ?"<<std::endl;
 		}
